data-extraction/requirements.txt
asyncio
crawl4ai
azure-storage-blob
python-dotenv
logging
requests
beautifulsoup4
scrapy
twisted
selenium
webdriver-manager
pytube
youtube-transcript-api